relocating functions to xlsutils
